requirements.txt
autopep8==1.4.4
certifi==2019.6.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.3
coveralls==1.8.1
dj-database-url==0.5.0
Django==2.0.3
django-auditlog==0.4.5
django-cors-headers==3.0.2
django-jsonfield==1.2.0
django-rest-swagger==2.2.0
djangorestframework==3.9.4
docopt==0.6.2
gunicorn==19.9.0
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
openapi-codec==1.3.2
pep8==1.7.1
psycopg2-binary==2.8.3
pycodestyle==2.5.0
python-dateutil==2.6.0
pytz==2019.1
requests==2.22.0
simplejson==3.16.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.25.3
